third_party/WebKit/LayoutTests/shadow-dom/crashes/slots-get-computed-style.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id= "host"></div>
<script>
test(() => {
  var root = host.attachShadow({ mode: 'open' });
  root.innerHTML = '<slot></slot>';
  assert_equals(getComputedStyle(root.querySelector('slot')).color, 'rgb(0, 0, 0)');
}, 'Getting CSS property value for a slot should not cause a crash');
</script>
